python-backend-engineer/pyproject.toml
[project]
name = "python-backend-engineer"
version = "0.1.0"
description = "Technical test for Python Backend Engineer candidates at Patrowl."
authors = [
    {name = "AntoineMenand",email = "antoine.menand@patrowl.io"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-stubs (>=5.2.5,<6.0.0)",
    "djangorestframework-stubs (>=3.16.3,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
